Изменил формулу расчета зп в таблице, начал добавлять удаленный доступ к железкам

+++ src/main/java/com/microel/trackerbackend/misc/ResponseWorkEstimationForm.java
@@ -52,7 +52,7 @@ public class ResponseWorkEstimationForm {
     @Builder
     public static class EmployeeRatioValue{
         private Float ratio;
-        private Integer sum;
+        private Float sum;
     }
 
 }
